Fix BCAO

+++ Distributed-Computing/scan.cuh
@@ -1,4 +1,4 @@
-long blockscan(int *output, int *input, int length, bool bcao);
+float blockscan(int *output, int *input, int length, bool bcao);
 float scan(int *output, int *input, int length, bool bcao);
 
 void scanLargeDeviceArray(int *output, int *input, int length, bool bcao);
